Hack to get delegations working by abusing the metadata.
According to the current Uptane spec, custom metadata should match. We
are not consistent with that, and this makes it even worse. If we find a
custom URI in a target from the images repo, but not in the target from
the director, use the URI from the images repo.

Note that by default right now, we ignore most custom metadata in the
images repo and only check and use what we get from the director.

Signed-off-by: Patrick Vacek <patrickvacek@gmail.com>